Fixed styling issue with to many notebooks
If there were to many notebooks, the layout was
bad and the list went out of the viewport.
